==> 1. Data Cleaning.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np

# Load Training Data

training_dataframe = pd.read_csv(filepath_or_buffer="./data/TrainingSet.csv")
print('Total number of training records',len(training_dataframe))

training_dataframe.head(10)

training_dataframe.describe()

# All the attributes have significant sd

np.array(training_dataframe.describe().loc['std'])<0.001

# Max number of Nones in a Columns

training_dataframe.isnull().sum().max()

# Columns having at least one None Value

null_column_series = training_dataframe.isnull().sum()>0
null_column_names = list(null_column_series[null_column_series].index)
print("Column Name -> None Counts :\n",training_dataframe[null_column_names].isnull().sum(),'\n')
print("Number of Columns having None : ",len(null_column_names),'\n')
print("Column Names containing at least one None value : \n",null_column_names)

# Number of rows having None Values

null_rows = training_dataframe.isnull().sum(axis=1)>0
print(len(null_rows[null_rows]))

feature_none_column_rows = training_dataframe.isnull().sum(axis=1)==31
print("Number of Rows having all 31 columns as None: ",len(feature_none_column_rows[feature_none_column_rows]))
decision_none_column_rows = training_dataframe.isnull().sum(axis=1)==1
print("Number of Rows having all 1 columns as None: ",len(decision_none_column_rows[decision_none_column_rows]))

# Missing data percentage wrt decision column

missing_percentage = len(decision_none_column_rows[decision_none_column_rows])*100/len(training_dataframe)
print("missing data percentage: %0.2f"%missing_percentage,"%")

non_null_rows = training_dataframe.isnull().sum(axis=1)==0
print(len(non_null_rows[non_null_rows]))
non_null_rows_index = non_null_rows[non_null_rows].index

# Select only those row for which 'Machine_State' has valid value

decision_states = ['Good','Bad']
decision_attribute_name = 'Machine_State'
training_dataframe_cleaned = training_dataframe[training_dataframe[decision_attribute_name].isin(decision_states)]

print("Original Training dataframe size : ",len(training_dataframe))
print("Cleaned Training dataframe size : ",len(training_dataframe_cleaned))
print("Number of rows dropped : ", len(training_dataframe)-len(training_dataframe_cleaned))

# Save the Cleaned Training Data

training_dataframe_cleaned.to_csv("./data/cleaned_training_data.csv",index=False)
