==> 4. Prediction.ipynb <==
# Import Libraries

import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn import datasets
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import SelectFromModel
from sklearn.metrics import accuracy_score
from sklearn.preprocessing import Imputer,LabelEncoder,StandardScaler
from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LogisticRegression
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis,QuadraticDiscriminantAnalysis
from sklearn.decomposition import PCA
from sklearn.neighbors import LocalOutlierFactor
from sklearn.model_selection import KFold
from sklearn.svm import NuSVC
import matplotlib.pyplot as plt
from sklearn.ensemble import AdaBoostClassifier
from sklearn.naive_bayes import GaussianNB
from imblearn.over_sampling import SMOTE, ADASYN
from sklearn.metrics import roc_curve
from sklearn.externals import joblib


%matplotlib inline 

# Load the Test Dataframe

test_dataframe = pd.read_csv("./data/TestingSet.csv")

test_data_matrix = test_dataframe.as_matrix()

X_test = test_data_matrix[:,1:]


# Load the Transformers and the Model

imputer = joblib.load('./data/model/imputer.pkl')
encoder = joblib.load('./data/model/encoder.pkl')
scaler = joblib.load('./data/model/scaler.pkl')
pca = joblib.load('./data/model/pca.pkl')
sfm = joblib.load('./data/model/sfm.pkl')
clf_RF = joblib.load('./data/model/rf.pkl') 

# Apply all the Transformers one by one

X = imputer.transform(X_test)
X = scaler.transform(X)
X = pca.transform(X)
X = sfm.transform(X)

# Apply the model on the data with Threshold = 0.561261261261

prediction = (clf_RF.predict_proba(X)[:,1]>0.561261261261).astype(np.int)

prediction

# Apply the Enocder to get the real labels back

predicted_labels = encoder.inverse_transform(prediction)

predicted_labels

# Prepare the Submission Dataframe and save it

submission_dataframe = pd.DataFrame()
submission_dataframe['Sl No.'] = [i for i in range(1,len(predicted_labels)+1)]
submission_dataframe['Machine_State'] = predicted_labels

submission_dataframe.head(5)

submission_dataframe.to_csv('./data/Submission.csv',index=False)
